separation of runtime and non-runtime libs
So far this only effects building ucommon itself, and really only on windows,
where we build ucommon as a runtime dll when "build runtime" is selected,
but we then build usecure as a static library.  Sipwitch builds it's
local library as a runtime only and has no extra libs that may go either
way.  Bayonne and apennine do the same.

The division between runtime and non-runtime libraries was originally
created because dll runtimes are required for servers that link shared library
code with plugins, but many of the supporting libraries that are missing
in windows which might get used (openssl, osip2, srtp, etc) probably will
get built static, and we do not want to "export" locally bound instances
into new dlls or servers when the same code may then get inserted into runtime
plugins.  In fact, sipwitch plugins explicitly use a back-call interface
into the server to make exosip related calls to avoid having to call exosip
functions directly and hence get static links of the same exosip2 and openssl
functions as the server.