commoncpp is shared or static, never runtime
There are issues with runtime mode and even dll builds of ccrtp itself,
and by extension commoncpp, on some platforms.  ccrtp should not be
runtime anyway since it may link to static libs on those platforms.  The
simplest way is to fix __EXPORT in commoncpp config.h rather than changing
everything in commoncpp and downstream to __SHARED.