removed separate commoncpp enable option
The --disable-commoncpp/--enable-commoncpp option is now gone from
configure.  It is handled directly with the --enable/disable
stdcpp option.  This was also done for CMakeLists BUILD_COMPAT
merged with BUILD_STDLIB.  The enable-commoncpp with disable-stdcpp
possibility was invalid, and this also simplifies configuration.
Also commoncpp now feels more part of the whole.